MVC (part 3) To understand whether you're heading in the right direction, you need to see the data. Therefore, 1. In the view package, create a UsersView class that implements View. It will display the list of users.
2. Create a controller field, and a corresponding setter, in UsersView.

3. Implement the refresh method:
3.1. Display the phrase "All users:".
3.2. Display all of the users in dataModel.
Indent each user with a tab.
3.3. Display the following visual separator at the end
===================================================

4. I'm already interested in seeing what we've got.
Add to UsersView a public void fireShowAllUsersEvent() method that emulates a client event.
Access the controller and call the appropriate method to display all users.

5. The Solution class will emulate a user. Open the Solution class, go to the red method,
and use IDEA's hot keys to create a usersView property (field).
All you need is a setter. If you created a getter, delete it.

6. Run main. Oops, nothing is displayed :(
This is because the server sent the data, and it was updated in the DataModel, but the View doesn't know anything about it.
The View can't update itself. The Controller does that.
Go to the controller and update the View's data.
Remember, the data is stored in the Model.

7. Run main. I now have this output:
All users:
    User{name='A', id=1, level=1}
    User{name='B', id=2, level=1}
===================================================
8. The AbstractJeans class must implement a toString() method whose output starts with name of the subclass.